[TuxOnIce] Make image prep variables longs.
Switch image preparation variables from ints to longs as a beginning
on support for larger memory and image sizes.

+++ kernel/power/tuxonice.h
@@ -187,7 +187,7 @@ extern char alt_resume_param[256];
 
 extern void copyback_post(void);
 extern int toi_hibernate(void);
-extern int extra_pd1_pages_used;
+extern long extra_pd1_pages_used;
 
 #define SECTOR_SIZE 512
